Refactor MDX configuration and plugins
- Consolidated MDX plugin imports into a single index file
- Separated MDX plugins into individual files for better modularity
- Updated Next.js and React type dependencies
- Simplified package configuration
- Enabled Turbopack for development
