+++ src/mdx/recma/mdx-annotations.mjs
@@ -1,0 +1,3 @@
+import { mdxAnnotations } from 'mdx-annotations'
+
+export default mdxAnnotations.recma
