+++ src/mdx/rehype/parse-code-blocks.mjs
@@ -1,0 +1,14 @@
+import { visit } from 'unist-util-visit'
+
+export default function rehypeParseCodeBlocks() {
+  return (tree) => {
+    visit(tree, 'element', (node, _nodeIndex, parentNode) => {
+      if (node.tagName === 'code' && node.properties.className) {
+        parentNode.properties.language = node.properties.className[0]?.replace(
+          /^language-/,
+          '',
+        )
+      }
+    })
+  }
+}
