+++ src/mdx/remark/mdx-annotations.mjs
@@ -1,0 +1,3 @@
+import { mdxAnnotations } from 'mdx-annotations'
+
+export default mdxAnnotations.remark
